+++ package.json
@@ -20,6 +20,7 @@
     "axios": "^1.5.0",
     "chart.js": "^4.4.0",
     "chartjs-plugin-datalabels": "^2.2.0",
+    "formik": "^2.4.5",
     "numeral": "^2.0.6",
     "react": "^18.2.0",
     "react-apexcharts": "^1.4.1",
@@ -34,13 +35,15 @@
     "react-scripts": "^5.0.1",
     "react-table": "^7.8.0",
     "recharts": "^2.8.0",
+    "recoil": "^0.7.7",
     "serve": "^14.2.1",
     "simplebar-react": "^3.2.4",
     "styled-components": "^6.0.4",
     "typeorm": "^0.3.17",
     "typescript": "^4.9.5",
     "victory": "^36.9.1",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "yup": "^1.3.3"
   },
   "scripts": {
     "start": "react-scripts start",
