+++ src/Routes.tsx
@@ -6,10 +6,15 @@ import LoginPage from './pages/login/LoginPage';
 import PreviousPage from './pages/archive/PreviousPage';
 import ArchiveDetailPage from './pages/archive/ArchiveDetailPage';
 // import SettingsPage from './pages/setting/SettingsPage';
-import SignUp1Page from './pages/signUp/SignUp1Page';
+import { SignUp1Page } from './pages/signUp/SignUp1Page';
+//import SignUp1Page from './pages/signUp/old/OldSignUp1Page';
 import SignUp2Page from './pages/signUp/SignUp2Page';
+//import SignUp2Page from './pages/signUp/old/OldSignUp2Page';
 import SignUp3Page from './pages/signUp/SignUp3Page';
+//import SignUp3Page from './pages/signUp/old/OldSignUp3Page';
+//import { SignUp4Page, SignUp4PageCandidate, SignUp4PagePasser } from './pages/signUp/old/OldSignUp4Page';
 import { SignUp4Page, SignUp4PageCandidate, SignUp4PagePasser } from './pages/signUp/SignUp4Page';
+
 import { SignUp5Page, SignUp5Complete } from './pages/signUp/SignUp5Page';
 import DeletePage from './pages/delete/DeletePage';
 //import DashboardMainPage from './admin/AdminPage';
